remove prevent-page-scroll mixing and move to component
