+++ addon/styles/emberui/eui-prevent-page-scroll.scss
@@ -21,6 +21,6 @@
 }
 
 .eui-pevent-page-scroll--net-content {
-  height: 10000px;
-  width: 10000px;
+  height: 100000px;
+  width: 100000px;
 }
